Pong!
Created a basic pong game, and fixed some bugs with the I/O input for keyboard events. The pong game manages to fit exactly into a single page (exactly 256 instructions). Hopefully, the game will get a few more things, but it'll be hard to simplify it even more to fit it all still in one page.